Support DERIVE types for Permon counters
I need to monitor some perfmon counters which just increment (e.g. emails sent from the SMTPSVC). This is what the DERIVE type is used for - but it was unsupported.

I've added support for specifying the .type in the ini file.

I also had to add support for changing the type according to CounterFormat because RRD doesn't support using floats for DERIVE types.